+++ src/main/java/org/realityforge/braincheck/BrainCheckTestUtil.java
@@ -1,7 +1,9 @@
 package org.realityforge.braincheck;
 
 import java.lang.reflect.Field;
+import java.util.function.Supplier;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 
 /**
  * Utility class for interacting with BrainCheck in tests.
@@ -11,6 +13,50 @@ public final class BrainCheckTestUtil
 {
   private BrainCheckTestUtil()
   {
+  }
+
+  /**
+   * The type of the guard.
+   * This is only used during development process and will be optimized out in production mode.
+   */
+  @GwtIncompatible
+  public enum GuardType
+  {
+    INVARIANT,
+    API_INVARIANT
+  }
+
+  /**
+   * Interface used to receive details about an invoked guard.
+   * This is only used internally during development to collect the guards/invariants/etc to ensure
+   * that they comply with patterns and to ensure they are documented appropriately.
+   */
+  @GwtIncompatible
+  public interface OnGuardListener
+  {
+    void onGuard( @Nonnull GuardType type, @Nonnull String message, @Nonnull StackTraceElement[] stackTrace );
+  }
+
+  /**
+   * Specify a callback that is invoked anytime the {@link Guards#invariant(Supplier, Supplier)} method or
+   * the {@link Guards#apiInvariant(Supplier, Supplier)} method is invoked while in development mode.
+   *
+   * @param onGuardListener the listener.
+   */
+  public static void setOnGuardListener( @Nullable final OnGuardListener onGuardListener )
+  {
+    if ( null == onGuardListener )
+    {
+      Guards.setOnGuardListener( null );
+    }
+    else
+    {
+      final Guards.OnGuardListener listener =
+        ( type, message, stackTrace ) -> onGuardListener.onGuard( GuardType.valueOf( type.name() ),
+                                                                  message,
+                                                                  stackTrace );
+      Guards.setOnGuardListener( listener );
+    }
   }
 
   /**
@@ -37,6 +83,7 @@ public final class BrainCheckTestUtil
       setCheckInvariants( true );
       setCheckApiInvariants( true );
     }
+    setOnGuardListener( null );
   }
 
   /**